+++ age/src/cli_common/recipients.rs
@@ -2,8 +2,7 @@ use std::io::{self, BufReader};
 
 use super::StdinGuard;
 use super::{identities::parse_identity_files, ReadError};
-use crate::identity::RecipientsAccumulator;
-use crate::{x25519, Recipient};
+use crate::{identity::RecipientsAccumulator, tag, tagpq, x25519, Recipient};
 
 #[cfg(feature = "plugin")]
 use crate::{cli_common::UiCallbacks, plugin};
@@ -56,6 +55,10 @@ fn parse_recipient(
     recipients: &mut RecipientsAccumulator,
 ) -> Result<(), ReadError> {
     if let Ok(pk) = s.parse::<x25519::Recipient>() {
+        recipients.push(Box::new(pk));
+    } else if let Ok(pk) = s.parse::<tag::Recipient>() {
+        recipients.push(Box::new(pk));
+    } else if let Ok(pk) = s.parse::<tagpq::Recipient>() {
         recipients.push(Box::new(pk));
     } else if let Some(pk) = {
         #[cfg(feature = "ssh")]