+++ age/src/native/x25519.rs
@@ -1,4 +1,4 @@
-//! The "x25519" recipient type, native to age.
+//! The classic recipient type, native to age.
 
 use std::collections::HashSet;
 use std::fmt;
@@ -26,10 +26,10 @@ const PUBLIC_KEY_PREFIX: &str = "age";
 pub(super) const X25519_RECIPIENT_TAG: &str = "X25519";
 const X25519_RECIPIENT_KEY_LABEL: &[u8] = b"age-encryption.org/v1/X25519";
 
-pub(super) const EPK_LEN_BYTES: usize = 32;
-pub(super) const ENCRYPTED_FILE_KEY_BYTES: usize = FILE_KEY_BYTES + 16;
+pub(crate) const EPK_LEN_BYTES: usize = 32;
+pub(crate) const ENCRYPTED_FILE_KEY_BYTES: usize = FILE_KEY_BYTES + 16;
 
-/// The standard age identity type, which can decrypt files encrypted to the corresponding
+/// The classic age identity type, which can decrypt files encrypted to the corresponding
 /// [`Recipient`].
 #[derive(Clone)]
 pub struct Identity(StaticSecret);
@@ -141,7 +141,7 @@ impl crate::Identity for Identity {
     }
 }
 
-/// The standard age recipient type. Files encrypted to this recipient can be decrypted
+/// The classic age recipient type. Files encrypted to this recipient can be decrypted
 /// with the corresponding [`Identity`].
 ///
 /// This recipient type is anonymous, in the sense that an attacker can't tell from the