+++ rage/CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+### Added
+- Support for the new native age recipient types:
+  - `age1tag1..`
+  - `age1tagpq1..`
 
 ### Changed
-- MSRV is now 1.70.0.
+- MSRV is now 1.74.0.
 
 ## [0.6.1, 0.7.2, 0.8.2, 0.9.3, 0.10.1, 0.11.1] - 2024-12-18
 ### Security